UDP Networking
NetworkMessage Options to create a message with custom UUID.
More Socket tests for making the client work - but can't get it to work locally.

##### Theory:
Local client does not get message because the server reads the message before the client can read it. Therefore, the client never gets the message and freezes.

Signed-off-by: Xtarii <carlalvinelias@gmail.com>